+++ Krowi_AchievementFilter/CHANGELOG.md
@@ -1,9 +1,16 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-## [Unreleased]
+## [9.0.2.8.0]
 ### Added
-- Battle for Azeroth zones
+- Battle for Azeroth zones (220+ achievements for Horde and Alliance combined)
+- The width of the categories list can be adjusted in options; negative values will narrow it, positive values will widen it
+
+### Changed
+- Interface => Options overhaul, using Ace3 for it now
+
+### Fixed
+- Issue #13: Updating addon in the past could lead to errors if new SavedVariables were added
 
 ## [v9.0.2.7.0] - 2021-01-09
 ### Added
@@ -75,6 +82,7 @@ All notable changes to this project will be documented in this file.
 - The need for almost all localization
 
 [Unreleased]: https://github.com/TheKrowi/Krowi_Addons/tree/master/Krowi_AchievementFilter
+[9.0.2.8.0]: https://github.com/TheKrowi/Krowi_Addons/tree/master/Krowi_AchievementFilter
 [v9.0.2.7.0]: https://www.curseforge.com/wow/addons/krowi-achievement-filter/files/3163711
 [90002.6.0]: https://www.curseforge.com/wow/addons/krowi-achievement-filter/files/3156244
 [90002.5.0]: https://www.curseforge.com/wow/addons/krowi-achievement-filter/files/3152510